chore: added __main__.py modules for tasks
