+++ src/argilla/tasks/users/__init__.py
@@ -12,20 +12,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-import typer
-
-from .create import create
-from .create_default import create_default
-from .migrate import migrate
-
-# from .create import create
-
-app = typer.Typer(help="Holds CLI commands for user and workspace management.", no_args_is_help=True)
-
-app.command(name="create_default", help="Creates default users and workspaces in the Argilla database.")(create_default)
-app.command(name="create", help="Creates a user and add it to the Argilla database.", no_args_is_help=True)(create)
-app.command(name="migrate")(migrate)
-
+from .__main__ import app
 
 if __name__ == "__main__":
     app()
